Add support for prefix on Android
